+++ ReleaseNotes.md
@@ -9,6 +9,9 @@
 * bugfix: 解决邮箱需要认证的情况下，邮件发送失败的问题
 * mysql: alert表增加字段: http_post_url
 * 文档: 增加query string简易教程
+* mysql: data_name表增加data_name字段唯一索引
+* bugfix: 解决dataname还在使用中，仍然可以删除的问题
+* elasticsearch监控增加avg,min,max,sum数值metric类型
 
 ## 0.1-RELEASE